backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
python-multipart
Pillow
reportlab
qrcode[pil]